python-fastapi-psycopg/requirements.txt
fastapi
uvicorn[standard]
psycopg
psycopg_pool
python-dotenv
black
